requirements.txt
click==6.6
Flask==0.11
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
uWSGI==2.0.13.1
Werkzeug==0.11.10
